user.controller.spec, service.spec test

+++ src/user/user.service.ts
@@ -1,18 +1,40 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { User } from './entity/user.entity';
 import { Repository } from 'typeorm';
+import * as bcrypt from 'bcrypt';
+import { envVariableKeys } from '../common/const/env.const';
+import { ConfigService } from '@nestjs/config';
 
 @Injectable()
 export class UserService {
   constructor(
     @InjectRepository(User)
     private readonly userRepository: Repository<User>,
+    private readonly configService: ConfigService,
   ) {}
   async create(createUserDto: CreateUserDto) {
-    return await this.userRepository.save(createUserDto);
+    const { email, password } = createUserDto;
+    const user = await this.userRepository.findOne({ where: { email: email } });
+
+    if (user) {
+      throw new BadRequestException('이미 가입한 이메일입니다.');
+    }
+
+    const hash = await bcrypt.hash(
+      password,
+      this.configService.get<number>(envVariableKeys.hashRounds),
+    );
+
+    await this.userRepository.save({ email, password: hash });
+
+    return this.userRepository.findOne({ where: { email: email } });
   }
 
   async findAll() {
@@ -28,16 +50,28 @@ export class UserService {
   }
 
   async update(id: number, updateUserDto: UpdateUserDto) {
+    const { password } = updateUserDto;
+
     const user = await this.userRepository.findOne({ where: { id: id } });
 
     if (!user) throw new NotFoundException('User not found');
 
-    await this.userRepository.update(id, updateUserDto);
+    const hash = await bcrypt.hash(
+      password,
+      this.configService.get<number>(envVariableKeys.hashRounds),
+    );
+    await this.userRepository.update(id, { ...updateUserDto, password: hash });
 
     return await this.userRepository.findOne({ where: { id: id } });
   }
 
-  remove(id: number) {
-    return this.userRepository.delete(id);
+  async remove(id: number) {
+    const user = await this.userRepository.findOne({ where: { id } });
+
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+    await this.userRepository.delete(id);
+    return id;
   }
 }
